requirements.txt
Flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.2.1
openai==2.7.2
PyPDF2==3.0.1